Move SerialNumberGenerator to book package and BorrowerIdGenerator to borrower package
