+++ src/test/java/seedu/address/logic/parser/CatalogParserTest.java
@@ -26,9 +26,9 @@ import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.ListCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Catalog;
-import seedu.address.model.SerialNumberGenerator;
 import seedu.address.model.book.Book;
 import seedu.address.model.book.SerialNumber;
+import seedu.address.model.book.SerialNumberGenerator;
 import seedu.address.model.book.TitleContainsKeywordPredicate;
 import seedu.address.testutil.BookBuilder;
 import seedu.address.testutil.BookUtil;
